slicedStationsのcustom hook化
